+++ cmake/SourcesAndHeaders.cmake
@@ -4,4 +4,6 @@ set(sources
 set(headers
     include/charybdis/chaos_audit.hpp
     include/charybdis/http_test_client.hpp
+    include/charybdis/safe_chaos.hpp
+    include/charybdis/subject_filter.hpp
     include/charybdis/test_helpers.hpp)